Fix - Move checker.utils tests into their dedicated unittest
It turns out there is not that much utils unittests.
